requirements.txt
pytest-playwright
ipython
nest_asyncio
